requirements.txt
unidecode
staticjinja
ruamel.yaml
six
